rename scootapi to cloud
