+++ cloud/README.md
@@ -1,0 +1,2 @@
+# Scoot Cloud #
+Scoot Cloud lets clients move data into Scoot, modify it, extract it, and perform operations on it.
