+++ cloud/api/README.md
@@ -1,0 +1,2 @@
+# Scoot Cloud API #
+Scoot Cloud offers a Thrift API.
