+++ cloud/api/client/cli.go
@@ -3,7 +3,7 @@ package client
 import (
 	"fmt"
 	"github.com/apache/thrift/lib/go/thrift"
-	"github.com/scootdev/scoot/scootapi/gen-go/scoot"
+	"github.com/scootdev/scoot/cloud/api/gen-go/scoot"
 	"github.com/spf13/cobra"
 	"log"
 )
