+++ cloud/api/client/run_job_cmd.go
@@ -2,7 +2,7 @@ package client
 
 import (
 	"fmt"
-	"github.com/scootdev/scoot/scootapi/gen-go/scoot"
+	"github.com/scootdev/scoot/cloud/api/gen-go/scoot"
 	"github.com/spf13/cobra"
 	"log"
 )
