+++ cloud/api/gen-go/scoot/constants.go
