+++ cloud/api/gen-go/scoot/proc.go
