+++ cloud/api/gen-go/scoot/ttypes.go
