+++ cloud/api/scoot.thrift
