+++ cloud/api/server/serve.go
@@ -3,9 +3,9 @@ package server
 import (
 	"fmt"
 	"github.com/apache/thrift/lib/go/thrift"
+	"github.com/scootdev/scoot/cloud/api/gen-go/scoot"
 	"github.com/scootdev/scoot/sched"
 	"github.com/scootdev/scoot/sched/queue"
-	"github.com/scootdev/scoot/scootapi/gen-go/scoot"
 	"log"
 	"time"
 )
