--- scootapi/api.go
@@ -1,4 +1,0 @@
-package scootapi
-
-// We should use go generate to run:
-//   thrift --gen go:package_prefix=github.com/scootdev/scoot/scootapi/gen-go/,thrift_import=github.com/apache/thrift/lib/go/thrift scoot.thrift
